+++ CHANGELOG.md
@@ -9,9 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Data Transfer Object classes generation from swagger spec
 
 ## [0.0.5] - 2020-05-08
+### Added
 - Validation subpackage based on rakit/validation
 - Request Validation Middleware
 - JsonApiResponseFactory
+- I18n support
+
+### Removed
+- Unused free-elephants/php-di dependency (but suggested) 
 
 ## [0.0.4] - 2020-04-27
 ### Fixed